Refactored scripts into a single script
