+++ src/scripts/readlineLogger.interface.ts
@@ -1,0 +1,3 @@
+export interface ReadlineLogger {
+    error: (message: string) => void;
+}
